some changes to facilitate the plugin disable
