+++ src/test/java/org/entando/entando/keycloak/services/OpenIDConnectServiceIntegrationTest.java
@@ -16,7 +16,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.entando.entando.keycloak.services.UserServiceIntegrationTest.activeUser;
+import static org.entando.entando.keycloak.services.UserManagerIntegrationTest.activeUser;
 
 public class OpenIDConnectServiceIntegrationTest {
 
@@ -27,7 +27,7 @@ public class OpenIDConnectServiceIntegrationTest {
     private IAuthorizationManager authorizationManager;
 
     private OpenIDConnectService oidcService;
-    private UserService userService;
+    private UserManager userManager;
 
     @Before
     public void setUp() {
@@ -35,27 +35,27 @@ public class OpenIDConnectServiceIntegrationTest {
         final KeycloakConfiguration configuration = KeycloakTestConfiguration.getConfiguration();
         final KeycloakService keycloakService = new KeycloakService(configuration);
         oidcService = new OpenIDConnectService(configuration);
-        userService = new UserService(authorizationManager, keycloakService, oidcService);
+        userManager = new UserManager(authorizationManager, keycloakService, oidcService);
 
         KeycloakTestConfiguration.deleteUsers();
     }
 
     @Test(expected = CredentialsExpiredException.class)
     public void testLoginWithExpiredCredentials() throws OidcException {
-        userService.addUser(activeUser(USERNAME, PASSWORD));
+        userManager.addUser(activeUser(USERNAME, PASSWORD));
         oidcService.login(USERNAME, PASSWORD);
     }
 
     @Test(expected = InvalidCredentialsException.class)
     public void testLoginWithInvalidCredentials() throws OidcException {
-        userService.addUser(activeUser(USERNAME, PASSWORD));
+        userManager.addUser(activeUser(USERNAME, PASSWORD));
         oidcService.login(USERNAME, "woodstock");
     }
 
     @Test
     public void testLoginSuccessfulAndTokenValidation() throws OidcException {
-        userService.addUser(activeUser(USERNAME, "woodstock"));
-        userService.updateUserPassword(USERNAME, PASSWORD);
+        userManager.addUser(activeUser(USERNAME, "woodstock"));
+        userManager.changePassword(USERNAME, PASSWORD);
         final AuthResponse response = oidcService.login(USERNAME, PASSWORD);
         assertThat(response.getAccessToken()).isNotEmpty();
